+++ cmd2/command_definition.py
@@ -2,8 +2,7 @@
 """
 Supports the definition of commands in separate classes to be composed into cmd2.Cmd
 """
-import functools
-from typing import Callable, Iterable, Optional, Type
+from typing import Optional, Type
 
 from .constants import COMMAND_FUNC_PREFIX
 from .exceptions import CommandSetRegistrationError
@@ -16,36 +15,6 @@ try:  # pragma: no cover
 
 except ImportError:   # pragma: no cover
     pass
-
-
-def _partial_passthru(func: Callable, *args, **kwargs) -> functools.partial:
-    """
-    Constructs a partial function that passes arguments through to the wrapped function.
-    Must construct a new type every time so that each wrapped function's __doc__ can be copied correctly.
-
-    :param func: wrapped function
-    :param args: positional arguments
-    :param kwargs: keyword arguments
-    :return: partial function that exposes attributes of wrapped function
-    """
-    def __getattr__(self, item):
-        return getattr(self.func, item)
-
-    def __setattr__(self, key, value):
-        return setattr(self.func, key, value)
-
-    def __dir__(self) -> Iterable[str]:
-        return dir(self.func)
-
-    passthru_type = type('PassthruPartial' + func.__name__,
-                         (functools.partial,),
-                         {
-                             '__getattr__': __getattr__,
-                             '__setattr__': __setattr__,
-                             '__dir__': __dir__,
-                         })
-    passthru_type.__doc__ = func.__doc__
-    return passthru_type(func, *args, **kwargs)
 
 
 def with_default_category(category: str):
@@ -78,8 +47,7 @@ class CommandSet(object):
 
     ``with_default_category`` can be used to apply a default category to all commands in the CommandSet.
 
-    ``do_``, ``help_``, and ``complete_`` functions differ only in that they're now required to accept
-    a reference to ``cmd2.Cmd`` as the first argument after self.
+    ``do_``, ``help_``, and ``complete_`` functions differ only in that self is the CommandSet instead of the cmd2 app
     """
 
     def __init__(self):
@@ -98,7 +66,7 @@ class CommandSet(object):
         else:
             raise CommandSetRegistrationError('This CommandSet has already been registered')
 
-    def on_unregister(self, cmd):
+    def on_unregister(self):
         """
         Called by ``cmd2.Cmd`` when a CommandSet is unregistered and removed.